remove Exp_ prefix in file name
