+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/GltfUtilities.h
