+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/TileContent.h
