+++ Cesium3DTilesSelection/src/BinaryToGltfConverter.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Cesium3DTilesSelection/Exp_GltfConverterResult.h>
+#include <Cesium3DTilesSelection/GltfConverterResult.h>
 #include <CesiumGltfReader/GltfReader.h>
 
 #include <gsl/span>
@@ -8,9 +8,13 @@
 #include <cstddef>
 
 namespace Cesium3DTilesSelection {
-struct CmptToGltfConverter {
+struct BinaryToGltfConverter {
+public:
   static GltfConverterResult convert(
-      const gsl::span<const std::byte>& cmptBinary,
+      const gsl::span<const std::byte>& gltfBinary,
       const CesiumGltfReader::GltfReaderOptions& options);
+
+private:
+  static CesiumGltfReader::GltfReader _gltfReader;
 };
 } // namespace Cesium3DTilesSelection
