--- Cesium3DTilesSelection/src/Exp_GltfConverterResult.cpp
@@ -1,1 +1,0 @@
-#include <Cesium3DTilesSelection/Exp_GltfConverterResult.h>
