--- Cesium3DTilesSelection/src/Exp_TilesetContentLoaderResult.cpp
@@ -1,1 +1,0 @@
-#include "Exp_TilesetContentLoaderResult.h"
