+++ Cesium3DTilesSelection/src/GltfConverterResult.cpp
@@ -1,0 +1,1 @@
+#include <Cesium3DTilesSelection/GltfConverterResult.h>
