+++ Cesium3DTilesSelection/src/GltfConverters.cpp
@@ -1,4 +1,4 @@
-#include <Cesium3DTilesSelection/Exp_GltfConverters.h>
+#include <Cesium3DTilesSelection/GltfConverters.h>
 
 #include <spdlog/spdlog.h>
