+++ Cesium3DTilesSelection/src/ImplicitOctreeLoader.h
@@ -1,27 +1,21 @@
 #pragma once
 
-#include "Exp_SubtreeAvailability.h"
-#include "Exp_TilesetContentLoader.h"
+#include "SubtreeAvailability.h"
+#include "TilesetContentLoader.h"
 
+#include <CesiumGeometry/OctreeTileID.h>
 #include <CesiumGeometry/OrientedBoundingBox.h>
 #include <CesiumGeospatial/BoundingRegion.h>
-#include <CesiumGeospatial/S2CellBoundingVolume.h>
-
-#include <cmath>
-#include <unordered_set>
-#include <variant>
-#include <vector>
 
 namespace Cesium3DTilesSelection {
-using ImplicitQuadtreeBoundingVolume = std::variant<
+using ImplicitOctreeBoundingVolume = std::variant<
     CesiumGeospatial::BoundingRegion,
-    CesiumGeospatial::S2CellBoundingVolume,
     CesiumGeometry::OrientedBoundingBox>;
 
-class ImplicitQuadtreeLoader : public TilesetContentLoader {
+class ImplicitOctreeLoader : public TilesetContentLoader {
 public:
   template <typename ImplicitBoundingVolumeType>
-  ImplicitQuadtreeLoader(
+  ImplicitOctreeLoader(
       const std::string& baseUrl,
       const std::string& contentUrlTemplate,
       const std::string& subtreeUrlTemplate,
@@ -51,24 +45,24 @@ public:
 
   uint32_t getAvailableLevels() const noexcept;
 
-  const ImplicitQuadtreeBoundingVolume& getBoundingVolume() const noexcept;
+  const ImplicitOctreeBoundingVolume& getBoundingVolume() const noexcept;
 
   void addSubtreeAvailability(
-      const CesiumGeometry::QuadtreeTileID& subtreeID,
+      const CesiumGeometry::OctreeTileID& subtreeID,
       SubtreeAvailability&& subtreeAvailability);
 
 private:
   static std::string resolveUrl(
       const std::string& baseUrl,
       const std::string& urlTemplate,
-      const CesiumGeometry::QuadtreeTileID& quadtreeID);
+      const CesiumGeometry::OctreeTileID& octreeID);
 
   std::string _baseUrl;
   std::string _contentUrlTemplate;
   std::string _subtreeUrlTemplate;
   uint32_t _subtreeLevels;
   uint32_t _availableLevels;
-  ImplicitQuadtreeBoundingVolume _boundingVolume;
+  ImplicitOctreeBoundingVolume _boundingVolume;
   std::vector<std::unordered_map<uint64_t, SubtreeAvailability>>
       _loadedSubtrees;
 };
