+++ Cesium3DTilesSelection/src/SubtreeAvailability.h
