+++ Cesium3DTilesSelection/src/TilesetContentLoaderResult.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "Exp_TilesetContentLoader.h"
+#include "TilesetContentLoader.h"
 
-#include <Cesium3DTilesSelection/Exp_ErrorList.h>
+#include <Cesium3DTilesSelection/ErrorList.h>
 #include <Cesium3DTilesSelection/Tile.h>
 #include <CesiumAsync/IAssetAccessor.h>
 #include <CesiumGeometry/Axis.h>
