+++ Cesium3DTilesSelection/src/TilesetContentManager.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "Exp_TilesetContentLoader.h"
+#include "TilesetContentLoader.h"
 
-#include <Cesium3DTilesSelection/Exp_TileContent.h>
+#include <Cesium3DTilesSelection/TileContent.h>
 #include <Cesium3DTilesSelection/Tile.h>
 #include <Cesium3DTilesSelection/TilesetExternals.h>
 #include <Cesium3DTilesSelection/TilesetOptions.h>
