+++ Cesium3DTilesSelection/src/TilesetJsonLoader.cpp
@@ -1,9 +1,9 @@
-#include "Exp_TilesetJsonLoader.h"
+#include "TilesetJsonLoader.h"
 
-#include "Exp_ImplicitOctreeLoader.h"
-#include "Exp_ImplicitQuadtreeLoader.h"
+#include "ImplicitOctreeLoader.h"
+#include "ImplicitQuadtreeLoader.h"
 
-#include <Cesium3DTilesSelection/Exp_GltfConverters.h>
+#include <Cesium3DTilesSelection/GltfConverters.h>
 #include <Cesium3DTilesSelection/TileID.h>
 #include <CesiumAsync/AsyncSystem.h>
 #include <CesiumAsync/IAssetResponse.h>
